Split settings tabs into separate files
