+++ src/frontend/containers/Settings/Advanced.tsx
@@ -1,0 +1,99 @@
+import { getThumbnailPath, isDirEmpty } from 'common/fs';
+import { runInAction } from 'mobx';
+import { observer } from 'mobx-react-lite';
+import React, { useEffect, useState } from 'react';
+import { RendererMessenger } from 'src/ipc/renderer';
+import { Button, ButtonGroup, IconSet } from 'widgets';
+import { useStore } from '../../contexts/StoreContext';
+import { moveThumbnailDir } from '../../image/ThumbnailGeneration';
+import { ClearDbButton } from '../ErrorBoundary';
+
+export const Advanced = observer(() => {
+  const { uiStore, fileStore } = useStore();
+  const thumbnailDirectory = uiStore.thumbnailDirectory;
+
+  const [defaultThumbnailDir, setDefaultThumbnailDir] = useState('');
+  useEffect(() => {
+    RendererMessenger.getDefaultThumbnailDirectory().then(setDefaultThumbnailDir);
+  }, []);
+
+  const changeThumbnailDirectory = async (newDir: string) => {
+    const oldDir = thumbnailDirectory;
+
+    // Move thumbnail files
+    await moveThumbnailDir(oldDir, newDir);
+    uiStore.setThumbnailDirectory(newDir);
+
+    // Reset thumbnail paths for those that already have one
+    runInAction(() => {
+      for (const f of fileStore.fileList) {
+        if (f.thumbnailPath && f.thumbnailPath !== f.absolutePath) {
+          f.setThumbnailPath(getThumbnailPath(f.absolutePath, newDir));
+        }
+      }
+    });
+  };
+
+  const browseThumbnailDirectory = async () => {
+    const { filePaths: dirs } = await RendererMessenger.showOpenDialog({
+      properties: ['openDirectory'],
+      defaultPath: thumbnailDirectory,
+    });
+
+    if (dirs.length === 0) {
+      return;
+    }
+    const newDir = dirs[0];
+
+    if (!(await isDirEmpty(newDir))) {
+      if (
+        window.confirm(
+          `The directory you picked is not empty. Allusion might delete any files inside of it. Do you still wish to pick this directory?\n\nYou picked: ${newDir}`,
+        )
+      ) {
+        changeThumbnailDirectory(newDir);
+      }
+    } else {
+      changeThumbnailDirectory(newDir);
+    }
+  };
+
+  return (
+    <>
+      <h2>Storage</h2>
+
+      {/* Todo: Add support to toggle this */}
+      {/* <Switch checked={true} onChange={() => alert('Not supported yet')} label="Generate thumbnails" /> */}
+      <fieldset>
+        <legend>Thumbnail Directory</legend>
+        <div className="input-file">
+          <input readOnly className="input input-file-value" value={thumbnailDirectory} />
+          <Button
+            styling="minimal"
+            icon={IconSet.FOLDER_CLOSE}
+            text="Browse"
+            onClick={browseThumbnailDirectory}
+          />
+          {defaultThumbnailDir && defaultThumbnailDir !== uiStore.thumbnailDirectory && (
+            <Button
+              icon={IconSet.RELOAD}
+              text="Reset"
+              onClick={() => changeThumbnailDirectory(defaultThumbnailDir)}
+            />
+          )}
+        </div>
+      </fieldset>
+
+      <h2>Development</h2>
+      <ButtonGroup>
+        <ClearDbButton />
+        <Button
+          onClick={RendererMessenger.toggleDevTools}
+          styling="outlined"
+          icon={IconSet.CHROME_DEVTOOLS}
+          text="Toggle DevTools"
+        />
+      </ButtonGroup>
+    </>
+  );
+});
