+++ src/frontend/containers/Settings/Shortcuts.tsx
@@ -8,6 +8,19 @@ import { defaultHotkeyMap, IHotkeyMap } from '../../stores/UiStore';
 import { camelCaseToSpaced } from 'common/fmt';
 import { Button, IconSet, keyComboToString } from 'widgets';
 
+export const Shortcuts = observer(() => {
+  return (
+    <>
+      <h2>Keyboard Shortcuts</h2>
+      <p>
+        Click on a key combination to modify it. After typing your new combination, press Enter to
+        confirm or Escape to cancel.
+      </p>
+      <HotkeyMapper />
+    </>
+  );
+});
+
 export const HotkeyMapper = observer(() => {
   const { uiStore } = useStore();
   const [changed, onChange] = useState<keyof IHotkeyMap | null>(null);
